+++ src/Data/Morpheus/Server/Deriving/Encode.hs
@@ -34,11 +34,10 @@ import Data.Maybe
     maybe,
   )
 import Data.Morpheus.Kind
-  ( ENUM,
-    GQL_KIND,
+  ( GQL_KIND,
     INTERFACE,
-    OUTPUT,
     SCALAR,
+    TYPE,
   )
 import Data.Morpheus.Server.Deriving.Channels
   ( ChannelsConstraint,
@@ -175,11 +174,7 @@ class EncodeKind (kind :: GQL_KIND) a o e (m :: * -> *) where
 instance (GQLScalar a, Monad m) => EncodeKind SCALAR a o e m where
   encodeKind = pure . ResScalar . serialize . unContextValue
 
--- ENUM
-instance EncodeConstraint o e m a => EncodeKind ENUM a o e m where
-  encodeKind = pure . exploreResolvers . unContextValue
-
-instance EncodeConstraint o e m a => EncodeKind OUTPUT a o e m where
+instance EncodeConstraint o e m a => EncodeKind TYPE a o e m where
   encodeKind = pure . exploreResolvers . unContextValue
 
 instance EncodeConstraint o e m a => EncodeKind INTERFACE a o e m where