+++ src/Data/Morpheus/Server/Types/GQLType.hs
@@ -12,7 +12,11 @@
 
 module Data.Morpheus.Server.Types.GQLType
   ( GQLType (..),
-    GQLTypeOptions (..),
+    GQLTypeOptions
+      ( fieldLabelModifier,
+        constructorTagModifier,
+        prefixTypeCategory
+      ),
     defaultTypeOptions,
     TypeData (..),
   )
@@ -20,6 +24,14 @@ where
 
 -- MORPHEUS
 import Data.Morpheus.Kind
+  ( GQL_KIND,
+    SCALAR,
+    TYPE,
+    ToValue,
+    WRAPPER,
+    isObject,
+    toValue,
+  )
 import Data.Morpheus.Server.Types.SchemaT
   ( SchemaT,
     TypeFingerprint (..),
@@ -69,14 +81,18 @@ data TypeData = TypeData
 
 data GQLTypeOptions = GQLTypeOptions
   { fieldLabelModifier :: String -> String,
-    constructorTagModifier :: String -> String
+    constructorTagModifier :: String -> String,
+    -- Type used as Input will be prefixed with: Input<TypeName>
+    -- type used as output will be prefixed with: Output<TypeName>
+    prefixTypeCategory :: Bool
   }
 
 defaultTypeOptions :: GQLTypeOptions
 defaultTypeOptions =
   GQLTypeOptions
     { fieldLabelModifier = id,
-      constructorTagModifier = id
+      constructorTagModifier = id,
+      prefixTypeCategory = False
     }
 
 getTypename :: Typeable a => f a -> TypeName
@@ -146,7 +162,7 @@ ignoreResolver (con, args) =
 --  @
 class ToValue (KIND a) => GQLType a where
   type KIND a :: GQL_KIND
-  type KIND a = OUTPUT
+  type KIND a = TYPE
 
   implements :: f a -> [SchemaT TypeName]
   implements _ = []
@@ -160,8 +176,8 @@ class ToValue (KIND a) => GQLType a where
   getDescriptions :: f a -> Map Text Description
   getDescriptions _ = mempty
 
-  typeOptions :: f a -> GQLTypeOptions
-  typeOptions _ = defaultTypeOptions
+  typeOptions :: f a -> GQLTypeOptions -> GQLTypeOptions
+  typeOptions _ = id
 
   getDirectives :: f a -> Map Text (Directives CONST)
   getDirectives _ = mempty