+++ test/Feature/Holistic/API.hs
@@ -54,7 +54,6 @@ import Prelude
     Show (..),
     String,
     const,
-    id,
   )
 
 data TestScalar